[DeckList] move SideboardPlan into separate file
