+++ libcockatrice_deck_list/CMakeLists.txt
@@ -9,7 +9,9 @@ set(HEADERS
     libcockatrice/deck_list/tree/inner_deck_list_node.h
     libcockatrice/deck_list/deck_list.h
     libcockatrice/deck_list/deck_list_history_manager.h
+    libcockatrice/deck_list/deck_list_node_tree.h
     libcockatrice/deck_list/deck_list_memento.h
+    libcockatrice/deck_list/sideboard_plan.h
 )
 
 if(Qt6_FOUND)
@@ -28,7 +30,7 @@ add_library(
   libcockatrice/deck_list/deck_list.cpp
   libcockatrice/deck_list/deck_list_history_manager.cpp
   libcockatrice/deck_list/deck_list_node_tree.cpp
-  libcockatrice/deck_list/deck_list_node_tree.h
+  libcockatrice/deck_list/sideboard_plan.cpp
 )
 
 add_dependencies(libcockatrice_deck_list libcockatrice_protocol)
